net-print/cups, cups-fuzz: Edit chromeos-version.sh
This allows the configure file to be removed and patches that update the
version number in configure.ac to be merged from upstream without
breaking builds.

Also for cups-fuzz, we prefer the default behavior of cros-workon
where it does not modify the version number when upreving the 9999
ebuild instead of basing it off the configure file.

BUG=none
TEST=cq

Change-Id: Ic43ce15773b486aedff5daea11afa558579577c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2765015
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: Pranav Batra <batrapranav@chromium.org>
Tested-by: Pranav Batra <batrapranav@chromium.org>
Auto-Submit: Pranav Batra <batrapranav@chromium.org>