net-misc/modemmanager-next: Allow MBIM bearer to access IPv4 address when requested IP type is IPv6
On T-mobile network when we request for IPv6 (only), to support roaming,
the network responds back with both IPv4 and IPv6 address.
MM (MBIM bearer) does not pass on the IPv4 address to shill,
which internally causes shill to connect/disconnect from the
network making the connection unusable.
This change basically passes the IPv4 config to shill if it's present.

BUG=b:183029202, b:162701578
TEST=emerge-hatch modemmanager-next && cros deploy && Check if we have
network access

Change-Id: I37bd136bda31152255075b433c845d40fa55622c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2802931
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Madhav . <madhavadas@google.com>
Tested-by: Madhav . <madhavadas@google.com>