Uprev protofiles and VERSION to Chromium 91.0.4467
This is required for usage in a TAST test which requires a Chrome Device
policy setting (which was introduced after the last uprev) to be set.

Also bump the protofiles version for policy-go-proto, which is used by
TAST, as well as authpolicy.

BUG=b:182835509
TEST=cros_run_unit_tests --board=volteer --packages authpolicy

Cq-Depend: chromium:2803981
Change-Id: I3cdcb919754f0b0aceaa281d56f2e719e26a8eaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2802933
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Prashant Malani <pmalani@chromium.org>