sys-fs/lxcfs: Introduce lxcfs 4.0.8
This is the latest upstream release on the 4.0.x LTS branch.

To allow for gradual rollouts, this version is installed at a
different path and has a different SLOT to the current version. This
allows for both versions to be installed on the same image.

This should eventually be moved back into portage-stable.

BUG=chromium:1097610
TEST=CQ

Cq-Depend: 2812823, 2813914
Disallow-Recycled-Builds: test-failures
Change-Id: Idb1f628a75e6a7ff0302ea2d273b93e5d0db839e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2813020
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>