app-emulation/lxc: Introduce lxc 4.0.9
This is the latest upstream release on the 4.0.x LTS branch.

To allow for gradual rollouts, this version is installed at a
different path and has a different SLOT to the current version. This
allows for both versions to be installed on the same image.

BUG=chromium:1097610
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: I0caf8f01bd0c73e6397219994d1ec739294982b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2813021
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>