app-emulation/lxd: Introduce lxd 4.0.6
This is the latest upstream release on the 4.0.x LTS branch.

To allow for gradual rollouts, this version is installed at a
different path and has a different SLOT to the current version. This
allows for both versions to be installed on the same image.

Note that tremplin depends on the client library shipped with LXD. As
this library uses the LXD REST API, and this API is supposed to be
backwards compatible, it should be possible for tremplin to continue
using the 3.17 client library with 4.0.6 LXD. This means we can have a
single tremplin build.

BUG=chromium:1097610
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: I12591daab01fec873b7a8245086c31962b65d5dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2813022
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>