sane-airscan: Add port in Host header
Upstream patch to add a quirk that forces the port to be included for
Epson printers.  This will be included in 0.99.26, but add it separately
in the meantime so that we can cherry pick it into M91 without pulling
in the whole ebuild upgrade.

BUG=b:183217222
TEST=Scan from ET-3750

Change-Id: I2bee57d966fb0df58e1341bbc484173895bd5e79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2811406
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Kyra Moed <kmoed@google.com>