binutils: add Gentoo patchset
When rebasing CrOS binutils the Gentoo patchset was forgotten.
Add it and refresh local patches on top.

The poison sysroot patch is dropped because the Gentoo archive
contains an improved version.

Because we're adding a src_prepare step, drop the hardcoded list
of PATCHES in favor of an eapply "${FILESDIR}" which is easier
to maintain.

BUG=chromium:1196871
TEST=Local builds; CQ

Change-Id: Ib300b320a326846ada5a0c2e7726e66187e84e37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2814379
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>