cros-kernel2.eclass: Remove kobject debugging from debug builds
This config is super noisy in the kernel logs as it basically enables a
handful of pr_debug() prints in the kernel's kobject code that we
practically never care to look at. It fills the kernel logs up rather
quickly so let's drop it.

BUG=None
TEST=USE="debug" emerge-$BOARD chromeos-kernel-5_4; Boot and see less
messages in dmesg

Change-Id: I6a1dde9a46151977af145b4c1462c69e980bc52a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2819293
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>