dev-util/turbostat: Build from upstream kernel
A single version of turbostat is used for all platforms. It does not
depend on the kernel version runnig on that platform. Therefore, it
does not depend on Chrome OS kernel sources.

The build from internal kernel sources was chosen to support private
modifications. The downside of this approach is that each change in the
internal kernel which affects the tool (for example include file changes)
triggers a test run on all Chromebooks in the lab, thus consuming
a substantial amount of lab resources.

To support changes on top of the upstream version of turbostat, it is not
necessary to build the tool from the internal kernel. Use the upstream
kernel instead, and apply private patches from patch files. Extract the
kernel version to use from the ebuild version.

BUG=chromium:1198329
TEST=Run turbostat

Change-Id: I4c3c1c306b7b4d1c408d0d70decb1bda3ff1086a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2823540
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>