sys-apps/util-linux: Upgrade to version 2.36.2
Drop our nosymfollow patch because it's merged upstream now and we've
ported that feature back to 3.18 kernels in CL:2644782 so the patching
conditional is not required anymore. We still keep this in chromium
overlay though because we add a patch to fix loopback device detection
(see https://github.com/karelzak/util-linux/issues/1202 that makes the
our umount calls in the build_image script work better) and we put back
python 3.6.

BUG=b:169246663
TEST=emerge-${BOARD} sys-apps/util-linux, cros deploy, run dmesg

Cq-Depend: chromium:2836630
Change-Id: I79db7dd9b2fc37ff19e12e58d487846fdb680814
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2426844
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>