(arc-)libdrm: Uprev to 2.4.105
radeonsi requires this in mesa 21.1

Set the CROS_WORKON_COMMIT a few commits after the upstream 2.4.105 tag
to include a revert for a patch that broke Chrome.

Also, make the non-9999 ebuild for arc-libdrm a symlink in the same
setup as libdrm to reduce duplication and avoid drift between the two
files.

BUG=b:185835037
TEST=Pass CQ dry run

Change-Id: I81b96dfe73a50a7dacbcd90cf6242774c59f69ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2837534
Tested-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Bas Nieuwenhuizen <basni@chromium.org>