qemu: stop statically linking system versions
We want to statically link the user programs so we can copy them into
the sysroots and run them standalone.  Since the emulation logic is
simple enough, this shouldn't be a problem.

For the system emulation, we need to stop statically linking it and
trying to use it directly outside the chroot.  This has never been
guaranteed to work -- statically linking glibc has long been a known
issue as it internally will dlopen modules whenever it needs to use
iconv functions or nss modules (which handle dns, passwd, etc...).
While it has mostly worked between versions, the latest glibc upgrade
broke it again, and it's not a bug in glibc as such.

Instead, we can use lddtree to create a copy of the x86_64 program
that we use outside the chroot, and it will take care of copying all
of its deps too.  This way we can run the program anywhere.

We do this in the qemu ebuild as the package is pulled down by the
simple chrome workflow, and so we don't have to maintain a separate
build step that runs whenever qemu itself updates.

BUG=chromium:1199436
TEST=running cros_vm works

Change-Id: Icf36d811ed3fd6681eafe74025877cf3ff4e85a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2843912
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>