zephyr-build-tools: Run tests on current source
The current approach of using 'pytest' appears to run zmake from the
chroot rather than the current source. This means that adding new tests
that rely on new zmake features is not possible.

Update it to use the current source in all cases.

BUG=b:185949077
TEST=FEATURES=test sudo -E emerge zephyr-build-tools
with crrev/c/2842209 applied
see that it passes

Change-Id: I32475f98ecaedffdfe4cfc4181f9ba46e9216d25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2844713
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>