dedede: Add pirika coreboot configuration
Create a new coreboot configuration for the pirika variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:184157747
TEST=FW_NAME=pirika emerge-dedede coreboot chromeos-bootimage
Ensure that image-pirika.*.bin are created

Change-Id: I00c5a192b95292c4c350426affbd5ebaeefee071
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2846154
Reviewed-by: Kirk Wang <kirk_wang@pegatron.corp-partner.google.com>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>
Reviewed-by: Paul2 Huang <paul2_huang@pegatron.corp-partner.google.com>
Tested-by: Kirk Wang <kirk_wang@pegatron.corp-partner.google.com>
Commit-Queue: Shou-Chieh Hsu <shouchieh@chromium.org>