servo-firmware: release v4p1 servo_v4p1_v2.0.8584+1a7e7e64c
This firmware is the same as the previous firmware, but it disables USB3
traffic to the DUT side. This is to guarantee a degree of stability as
the device is being rolled out, and we can root cause the usb3
instability.

BUG=b:186874604
TEST=sudo emerge servo-firmware
     sudo servo_updater -b servo_v4p1 -p
board: servo_v4p1
channel: stable
firmware: servo_v4p1_v2.0.8584+1a7e7e64c

Change-Id: I9c32369844206081a11706d93b942d9a4653f0b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2867831
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>