cros-rust: Add CROS_RUST=1 in ecargo_build
Intended use case:
- Crate A generates sources when it is emerged from input files
  that are only accessible when it emerges.
- Crate B depends on crate A, and this is reflected in the
  ebuild for crate B.
(Examples: cros-dbus-bindings or bindgen for *-sys)

The following scenarios are supported and need to work:
- local `cargo build` for crate A
- local `cargo build` for crate B
- emerge A
- emerge B

Add CROS_RUST environment variable to support the `emerge B`
case, since crate B can't access pre-generated source
in emerge, the build.rs script for crate A will skip the
source generation if both of the following are true:
- The generated source exists
- `CROS_RUST=1`

In local `cargo build` in both A and B:
if the generated file exists, we always remove
the file first and regenerate it.

In `emerge A`:
- Regenerate the source since the generated source not exists
- Deploy its source with generated sources to cros_rust_registry

In `emerge B`:
- Since B depends on A, A's source with the generated file is already in
  cros_rust_registry
- build.rs in A will be triggered in B's build but it'll skip source generation
  by checking CROS_RUST=1

Note: cros-dbus-bindings currently assumes PROFILE=release is only used
in emerge phase and use it to skip source generation in build.rs.
But it's not general enough since we could use PROFILE=release in local
crate build.

BUG=b:186004820
TEST=`cargo build` in cras-sys
TEST=`cargo build` in libcras
TEST=emerge-${BOARD} -j cras-sys
TEST=emerge-${BOARD} -j cras-sys libcras

Cq-Depend: chromium:2866787
Change-Id: Ie48ef828cb84365ec5d162292d709d654b035d58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2868445
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>