aosp-frameworks-ml-nn: allow NDEBUG to be set
NDEBUG can't be set for test targets, but is required for the core
library, so we're handling that in the BUILD.gn file now rather than in
the ebuild to be more specific as to where it needs to be unset.

See http://crrev.com/c/2870225.

BUG=b:186996014
TEST=FEATURES=test USE=-cros-debug emerge-volteer aosp-frameworks-ml-nn

Cq-Depend: chromium:2870225
Change-Id: Ice1bbc69a49727ecadaba266cf57fcfdf6d0a7d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2870206
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>