volteer: Add chronicler coreboot configuration
Create a new coreboot configuration for the chronicler variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:187318819
TEST=FW_NAME=chronicler emerge-volteer coreboot chromeos-bootimage
Ensure that image-chronicler.*.bin are created

Cq-Depend: chrome-internal:3833056
Change-Id: I451e1fe7e8c973394f9e69a959853711c2ff6340
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2878252
Tested-by: Sheng-Liang Pan <sheng-liang.pan@quanta.corp-partner.google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>