eclass/cros-rust: add support for envvar EXTRA_RUSTFLAGS
With this CL, we expose the envvar EXTRA_RUSTFLAGS to the build system,
on top of the already defined ebuild-specific local rustflags. This lets
us define extra flags by running the emerge command directly from chroot
that will then be passed to the packages like a normal cargo build
compilation would be expected to work.

BUG=b:181104775
TEST=`EXTRA_RUSTFLAGS="flag" emerge-$BOARD crosvm` and verified "flag" works

Change-Id: I75686d51a91ae19491b8df06f3624c336fd5e7e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2891871
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Morg <morg@chromium.org>
Tested-by: Morg <morg@chromium.org>