swap-init: replace blkdiscard command with dd
With util-linux version 2.36.2 blkdiscard is not working on the
dm-integrity device created by cryptsetup --integrity-no-wipe
because it is now probing the device, trying to read for signatures,
even with the force flag. With --integrity-no-wipe reads from sectors
that haven't been written to return an error. Replace blkdiscard -z
with dd if=/dev/zero to avoid this blkdiscard error.

BUG=b:187889770, b:184088189
TEST=emerge, deploy, swapon

Change-Id: Ie6c62888337b3d332b538518614fb5b59dd581ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2893498
Reviewed-by: Marcus Boerger <mboerger@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>