ti50-sdk: clarify logging output in build scripts
Our llvm-next testing failed to build ti50-sdk; the failure was in Rust
code, but all of the warnings/errors dumped to stdout & stderr from this
script made that unclear.

This CL silences output from clang as long as the build passes/fails
as expected. When the successfulness of the build doesn't meet
expectations, we'll dump stdout/stderr and break.

This makes no change to the built artifacts, so no revbump is included.

BUG=None
TEST=`emerge ti50-sdk` with some failures sprinkled in & without

Change-Id: I2cbcf89cc6ea945d4f7a8d666005108c4e2d68cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2906915
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>