sys-kernel/linux-headers: backport faccesat2 syscall definitions
This backport is required because newer glibc >= 2.33 will try
to call the newly added syscall (>= v5.8) and if -ENOSYS is
returned then it will fallback to the old faccesat syscall.

Minijail which is built against this linux-headers version needs
to be aware of the existence of the syscall otherwise it will
either just block it or fail at policy load due to unknown call.

BUG=b:187795855
TEST=Local builds; CQ.

Change-Id: I15f64f35d913eeecf4533e05f1d544a95d89486a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2917160
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>