brya: Add primus coreboot configuration
Create a new coreboot configuration for the primus variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:188272162
TEST=FW_NAME=primus emerge-brya coreboot chromeos-bootimage
Ensure that image-primus.*.bin are created

Cq-Depend: chromium:2940509
Change-Id: I34eb812d0278958280dbf99a32b1e705e334dc59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2932279
Tested-by: Scott Chao <scott_chao@wistron.corp-partner.google.com>
Tested-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>