ambassador: Add scout coreboot configuration
Create a new coreboot configuration for the scout variant
of the puff reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=None
TEST=FW_NAME=scout emerge-ambassador coreboot chromeos-bootimage
Ensure that image-scout.*.bin are created

Cq-Depend: chromium:2959844
Change-Id: Ic111df4bf182a569deae600087bdbd844f1f6821
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2938682
Tested-by: Jeff Chase <jnchase@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Jeff Chase <jnchase@google.com>