chromeos-base/percetto: Disable werror
It appears that percetto tries to enable a bunch of warnings as
errors, but cannot actually be built in this way because meson
versions prior to 0.56.0 would silently ignore the warning_level
option in subprojects. We are now updating meson, so this breaks
percetto. Fix by adding "-Dwerror=false" to emesonargs.

BUG=chromium:1215974
TEST=CQ

Change-Id: I7e3ff0e227a2771885c4e1c2d57d3ae8f58ff48e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2939413
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: John Bates <jbates@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>