chromeos-base/vm_*_tools: Make CROS_WORKON_SUBTREE finer grained
This variable is used by the build system to determine if a package
needs to be rebuilt after a change or not. Currently both
vm_guest_tools and vm_host_tools pull in the whole of vm_tools, which
means even changes that only affect guest tools force a rebuild and
test of almost all boards.

This CL changes both ebuilds to listing the subdirectories of vm_tools
they make use of. This should mean that e.g. changes to sommelier no
longer trigger rebuilds of host boards, and changes to concierge no
longer trigger rebuilds of termina (though this is of less value).

BUG=none
TEST=CQ

Cq-Depend: 2939113
Change-Id: Ia447c23583e7269efcb9059318618685f22f5b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2939416
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>