LLVM-next: use -z,nostart-stop-gc by default.
This seems to be only useful for instrumented profiling which we dont
use. It causes several packages to fail.
I may work on enabling this but it seems low priority.
This change only adds it to llvm-next.
Will be enabled with next compiler roll.

BUG=b:187743526
TEST=CQ

Change-Id: I01caff64078a6ed67a1cdc571bfb05f2c64b7b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2942906
Tested-by: Luis Lozano <llozano@chromium.org>
Auto-Submit: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>