brya: Add gimble coreboot configuration
Create a new coreboot configuration for the gimble variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:190334274
TEST=FW_NAME=gimble emerge-brya coreboot chromeos-bootimage
Ensure that image-gimble.*.bin are created

Cq-Depend: chromium:2960788
Change-Id: I46054a28b0063b7f511eb84150d4d309aa7a42ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2944516
Tested-by: Mark Hsieh <mark_hsieh@wistron.corp-partner.google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>