sys-libs/glibc: Backport fix for shm path autotest failure
An autotest started failing due to a dubious glibc codepath
which got removed after glibc 2.33, so we backport the relevant
patch to also remove it in 2.32.

This fix is speculative because the failure happens sporadically
and is hard to reproduce, likely a result of optimizations done
by the compiler, but removing the codepath should fix it.

BUG=chromium:1214312, b:190711694
TEST=Local builds & boot on EVE and Kevin, CQ.

Change-Id: I2a38ef96da2ef13b3b996ccb27b508ad5082d21b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2949488
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>