dev-libs/libxmlb: uprev to ver 0.3.2
This CL adds and modifies the ebuild file for the newest
version of libxmlb. Specifically, the PYTHON COMPAT and the
keywords are editted on the libxmlb-0.3.2.ebuild file so that the newest version
can work with the right boards. The CL also edits the Manifest to include only
information about the latest version. The edits are compatible with the
three boards that were tested (amd64, arm64, and arm).

There is also an additional symlink file called "lib-0.3.2-r1.ebuild"
that points to the new ebuild file.

BUG=b:189350744
TEST=Using emerge to build amd64, arm, and arm64 using the newest version of
libxmlb.

Disallow-Recycled-Builds: test-failures
Change-Id: Ie05744b531b7e0a242b20f229b9a59b241a29254
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2945132
Reviewed-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Tested-by: Daniel Campello <campello@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>