(2 of 2) Moving to src tree from portage
Keeping this in the overlay adds no value since it's not invoked by
portage.

This makes cq coverage rules worse and also makes it harder to share
logic across docker builds.

Initially, this was done to keep the build config close to the ebuild
that actually built the binary (for discoverability, but not worth the
tradeoff).

BUG=None
TEST=./build-dockerimage.sh

Change-Id: Ia5de8bc5f603b0b9f8487449c95e7b9bd889bac0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2966096
Tested-by: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>