net-misc/modemmanager-next: Set udev rule for PCIe modem mbim endpoint
This change sets the correct permissions for the new /dev endpoints
exposed by the CCCI driver for PCIe modems.

The change was tested by verifying permissions with

'ls -lah /dev/ttyC*'

and confirming that the required MBIM port belongs to the modem group
and received correct 660 permissions, and that the other /dev/ttyC*
endpoints exposed by the driver do not.

BUG=b:190145164
TEST=Verified endpoint permissions around reboot

Change-Id: I4d82dcae4e73e8c5190c71ae920dc1464fb1f83f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2976680
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>