brya: Add redrix coreboot configuration
Create a new coreboot configuration for the redrix variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:192052098
TEST=FW_NAME=redrix emerge-brya coreboot chromeos-bootimage
Ensure that image-redrix.*.bin are created

Cq-Depend: chromium:3016184
Change-Id: I4ae03923c1b2f482e4991457a579223c69f16cc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2992352
Tested-by: Wisley Chen <wisley.chen@quanta.corp-partner.google.com>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>