dnsmasq: upgrade 2.78 -> 2.85
The main reason for this upgrade is that dnsmasq has some
bugs not including the proper kernel headers which were
uncovered while upgrading from linux-headers 4.14 -> 5.10.

Aside from those fixes, the 2.78 release is quite old and
we can drop 3 of the downstream patches which are included
in 2.85, besides getting the the latest Gentoo ebuild fixes.

BUG=b:189218019
TEST=Local builds; CQ.

Change-Id: I9aabdf66d32c8294bff0f53b0f51f755fd93420d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2988404
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>