keeby: Add cappy2 coreboot configuration
Create a new coreboot configuration for the cappy2 variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:192035460
TEST=FW_NAME=cappy2 emerge-keeby coreboot chromeos-bootimage
Ensure that image-cappy2.*.bin are created

Cq-Depend: chromium:2992060
Change-Id: I0126b725f0ebf6af92d2e2d0dde30da97324c56d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2992357
Tested-by: Sunwei Li <lisunwei@huaqin.corp-partner.google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Henry Sun <henrysun@google.com>