delete mst_hub_updater
This package is the Realtek MST hub updater, which was used only for
testing and debugging of new support in flashrom in fwupd. It can now be
deleted because support for the device has been implemented in both
flashrom and fwupd.

BUG=b:192415024
TEST=None

Change-Id: I57d81447b01f39c389c708c8b385a08ad3a35007
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2999437
Tested-by: Peter Marheine <pmarheine@chromium.org>
Auto-Submit: Peter Marheine <pmarheine@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>