reland: sys-devel/binutils: bump 2.35.2 -> 2.36.1
This is a rebase on top of 2.36.1 in order to get the Intel
Keylocker assembler instructions support added in commit
c4694f172b51a21.

Some conflicts were due to the binutils internal hash table
API changing but nothing major or scary, just converted the
dwarf5 patches to use the new API and the rest of conflicts
were trivial.

The following patch was converted to proper patch format:
0015-gold-removes-iovec.patch

BUG=b:197038743
TEST=Local builds; CQ.

Change-Id: Idc10689a36019ae49dd7b676c9e41e5ba98a9403
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3198906
Reviewed-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>