Deprecate USE flag diag_payload and minidiag
For diag_payload:
It is for a diagnostic tool which based on legacy UI and deprecated.

For minidiag:
Mini-diag is set default enabled for newer devices, and we could still
use kernel secdata in runtime if we decide to disable it on certain
platforms.
Deprecate CONFIG_DIAGNOSTIC_UI in depthcharge and bmpblk, and set
them as default enabled.

In conclusion, deprecating these diagnostic related USE flag will not
affect newer devices.

BUG=b:190796342, b:181931817
TEST=build depthcharge
TEST=build chromeos-bmpblk
TEST=build chromeos-bootimage on volteer

Cq-Depend: chromium:3004223, chromium:3004656
Change-Id: I0d715aba63b4e05840961ab6b388e815ee40d1b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3005643
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Tested-by: Hsuan Ting Chen <roccochen@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>