brya: Add kano coreboot configuration
Create a new coreboot configuration for the kano variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:193052432
TEST=FW_NAME=kano emerge-brya coreboot chromeos-bootimage
Ensure that image-kano.*.bin are created

Cq-Depend: chromium:3018212
Change-Id: If7acd88d5b8272bc95456ef49ec357bda9dbdd0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3012160
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>