net-print/cups-filters: Restore TotalPageCount for PWG-Raster
Sets job-impressions to the correct page count so that the
value can be used to embed TotalPageCount in PWG-Raster
output. However, job-impressions is only embedded in the
document for PWG-Raster content (as it's not a feature for
CUPS-Raster).

BUG=b:172225466
TEST=Verify that CUPS Raster outputs did not change and
TotalPageCount is set in PWG-Raster output

Cq-Depend: chromium:3017607
Change-Id: I55eac3b9d20779c5dcdd975e5c97ff6e67ce3c90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3018498
Commit-Queue: Sean Kau <skau@chromium.org>
Tested-by: Sean Kau <skau@chromium.org>
Reviewed-by: Pranav Batra <batrapranav@chromium.org>