crosvm: enable libvda feature
Libvda is now an optional backend for accelerated video on crosvm and
needs to be mentioned explicitly as a feature.

This CL should only be merged when crosvm is upreved to a version that
recognizes the "libvda" feature (i.e. with crrev.com/c/3026351),
otherwise building will fail.

BUG=b:169295147
TEST=emerge-hatch chromeos-base/crosvm succeeds.

Change-Id: I38c6d8eccf700ff7816d9b40d71db96d23b453be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3023696
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>