chromeos-base/hps-firmware: Add ebuild for stage0 and co
The following is able to build the HPS stage 0 MCU rom
image. A stage 1 image can be built as well however the
inlining of the git hash is currently broken. This is
to be resolved in a follow up. Further, the sign-rom
tool is also built.

BUG=b:191820715,b:195612000
TEST=emerged stage0 binary.

Cq-Depend: chrome-internal:4049408, chrome-internal:4117572
Change-Id: Idb776e74b4059b4f3eeb58b4fe494c0f4c886ec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3115795
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Dan Callaghan <dcallagh@chromium.org>