Make test-services build for autotest-all
autotest-all is a special package that is set as a top package build
target for CI recipe workflows.

This leverages that existing CI integration to force a build dep on
test-services, so that they are built any time autotest-all is required.

Eventually, autotest-all will likely be renamed to something more
general, but starting here since the CI support is already present.

BUG=None
TEST=cq

Change-Id: I83afa11080210df487a195567af7d2857f17bed9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3020627
Commit-Queue: C Shapiro <shapiroc@chromium.org>
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Reviewed-by: Sean McAllister <smcallis@google.com>