sys-apps/fwupd: move from system-services to boot-services
This change allows for the daemon to be run before system-services is
started and fixes a regression that caused fwupd-at-boot to stop
working. This change does not make fwupd service to block boot but it
does delay its shutdown if it is indeed running.

BUG=b:191912505,b:193421830
TEST=reboot with pending update

Change-Id: I032565d3720deadb21d6602b91d616af6da7d20e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3022122
Tested-by: Daniel Campello <campello@chromium.org>
Auto-Submit: Daniel Campello <campello@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Daniel Campello <campello@chromium.org>