sys-apps/fwupd: Convert fwuptool-update to multi-instance job
This change allows for multiple concurrent instances of the
fwupdtool-update upstart task. These multiple instances will be issuing
concurrent operations to the fwupd daemon that will be in charge of
serializing them properly.

This change would introduce issues with multiple fwupdtool instances
(when fwupdmgr is unused and the daemon is not available). These
concurrent instances will be handled with a file lock in
/var/run/fwupdtool.

It remains to be address concurrent fwupdtool/fwupdmgr instances. This
might be handled with the same file lock but it remain orthogonal to
this change.

BUG=b:172985218
TEST=initctl emit fwupdtool-update

Change-Id: If7f593494a2f8cd42d596fffbf392a753d31a856
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3022124
Tested-by: Daniel Campello <campello@chromium.org>
Auto-Submit: Daniel Campello <campello@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Daniel Campello <campello@chromium.org>