wpa_supplicant-2_9: drop seccomp fallback
We've had a release cycle with seccomp enabled in a "run once" manner
(where it disables on the second run per-boot, to mitigate issues around
incomplete seccomp filters). This has a downside of providing
essentially no security benefit, because all an attacker has to do is
force one crash and then the sandbox is lightened.

Drop the fallback, and run with seccomp unconditionally.

BUG=b:177454191
TEST=wifi.*

Cq-Depend: chromium:3033682
Disallow-Recycled-Builds: all
Change-Id: I335b653b8f402cba276cce9ff68ecc701162ca30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3032766
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>