brya: Add taeko coreboot configuration
Create a new coreboot configuration for the taeko variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:193685558
TEST=FW_NAME=taeko emerge-brya coreboot chromeos-bootimage
Ensure that image-taeko.*.bin are created

Cq-Depend: chromium:3044270
Change-Id: Id63e1ed89f68499b14109ca94064d10ec8238483
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3044278
Reviewed-by: YH Lin <yueherngl@chromium.org>
Tested-by: Joey Peng <joey.peng@lcfc.corp-partner.google.com>
Tested-by: Kevin Chang <kevin.chang@lcfc.corp-partner.google.com>
Commit-Queue: YH Lin <yueherngl@chromium.org>