sys-apps/fwupd: Configuring fwupd to stop when powerd stops
A phrase was added to fwupd.conf which stops fwupd in the event that
powed is stopped. This is so that it can re-assess the system and
realize that powerd is absent, which it only does on startup.

BUG=b:194427863
TEST= Run Tast test developed here: crrev/c/3041173 and confirm that
powerd plugin is disabled after running "stop powerd"

Change-Id: I4ac1eff9e4fb3bce4390c3d1c2b39591656a3730
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3045776
Tested-by: George Popoola <gpopoola@google.com>
Reviewed-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Daniel Campello <campello@chromium.org>