chromeos-chrome: Explicitly set dcheck_always_on to true/false.
Currently, DCHECKs default to false for CrOS:
https://source.chromium.org/chromium/chromium/src/+/main:build/config/dcheck_always_on.gni;drc=4da74fd850f7abbfe249b5f208c1b370535d3b3e;l=32

Ideally we remove the "!is_chromeos" bit and let it default to true
on CrOS for developers. But we wouldn't necessarily want bots to
pick up a default of true (eg: sanitizer bots that set is_official
to false) when we don't want them to.

So this will explicitly set the arg to true or false for all emerge's
of chrome. We can then let it pick the default in simplechrome for
developers.

BUG=none
TEST=letting the CQ do the work

Change-Id: If8088d1b320f1990a287c96ed07d560f0be79914
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3057481
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Mike Wiitala <mwiitala@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>