arc-build.eclass: Explicitly bar meson from running binaries
arc-build-create-cross-file has a redundant copy of the code in
meson.eclass for generating cross-files. This is unnecessary because
multiple cross-files are merged together by meson. This means we can
strip out most of the code and only include the overrides.

This was originally done to set "needs_exe_wrapper = true" to prevent
meson from trying to run cross-build binaries during the build. This
is now done unconditionally upstream, but I am also adding a wrapper
script to allow unit tests to run for other packages. This needs to be
overridden for arc-build as the usual script doesn't handle ARC
binaries.

For now, just set "exe_wrapper = ''" which maintains the status quo
here that cross-built binaries cannot be run at build time.

BUG=chromium:1227738
TEST=emerge-zork-kernelnext media-libs/arc-mesa-amd

Change-Id: Id246304cd8e5f8aea090818c667cc1821e61f474
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3058026
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>