shill: change supplicant's default sae_pwe config
By default, supplicant set its sae_pwe config to 0 (hunting-and-pecking
only), but recent standards require H2E (hash-to-element) to be enabled.

This CL changes the configuration to let wpa_supplicant use both
hunting-and-pecking and hash-to-element, if wpa_supplicant is recent
enough to support that setting (July 2021 upstream rebase or newer).
wpa_supplicant's documentation mentions that the default value could be
changed in the future, when that happens we can revert this CL.

BUG=b:190725443
TEST=emerge-${BOARD} shill
TEST=grep sae_pwe /build/${BOARD}/usr/lib64/shill/shims/wpa_supplicant.conf

Change-Id: I2404cd9d0c34dda9169e4793ea55705e0d7eb174
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3059879
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>