Enable document scanning by default
With the use flag, we are going to install document scanning .so file
into rootfs. The .so file is about 8.1 MB. Please refer to the design
doc for more details:
go/cros-camera:dd:document-scanning

And according to the discussion in CrOS Small Council via the design
sketch:
go/cca-document-scanning-designsketch

The binary size is considered big for devices with 2GB rootfs.
Therefore, this CL enables document scanning by default excepts for
those with "embedded" or "minimal" feature, and having another CL to
exclude all 2GB rootfs devices:
https://crrev.com/c/3060133

BUG=b:180564352
TEST=None

Change-Id: Ic5fa0ae4b24944ceb2129d4e400b034dfbd97a66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3056664
Tested-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>