dedede: Add bugzzy coreboot configuration
Create a new coreboot configuration for the bugzzy variant
of the waddledoo reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:192521391
TEST=FW_NAME=bugzzy emerge-dedede coreboot chromeos-bootimage
Ensure that image-bugzzy.*.bin are created

Cq-Depend: chromium:3075129
Change-Id: I8aee24615ad95661b0c1a2fac4d13913dae3b5d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3066983
Reviewed-by: Raymond Chung <raymondchung@ami.corp-partner.google.com>
Reviewed-by: Henry Sun <henrysun@google.com>
Tested-by: Raymond Chung <raymondchung@ami.corp-partner.google.com>
Commit-Queue: Henry Sun <henrysun@google.com>