perfetto: disable warning about reserved-identifiers.
perfetto declares some ASAN identifiers so this warning
is not valuable in this case.

This change is in preparation for next compiler roll.
Current compiler does not support -Wreserved-identifier.
But this will not cause any problems because we are also
using -Wno-unknown-warning-option by default.

BUG=b:194728620
TEST=CQ

Change-Id: I08b632ccfa6767e91e97470825a4486973a30c9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3068514
Tested-by: Luis Lozano <llozano@chromium.org>
Auto-Submit: Luis Lozano <llozano@chromium.org>
Reviewed-by: Chinglin Yu <chinglinyu@chromium.org>